@@ -1,1 +1,1 @@
-2.7.2
+3.0.0
